fixed BYTEMAN-210
implemented support for array length accesses i.e. expressions of the
form arrayExpression.length

modified the type checker so AssignableExpressions know whether they
are type checkign as an LVALUE or RVALUE -- this i sneeded to ensure
attempts to update arrayExpression.length are rejected with a type
error

modifed the grammar so that array indexed terms can appear as the target
of a field dereference without needing bracketing

extended the test case in TestArray to check that array length
accesses are ok. the test case does not include an example where
arrayExpression.length is used as an LVALUE because this just throws
an exception but this was checked manually.