completed refactoring of check code fixes for BYTEMAN-190
Added ability to store info messages to RuleCheckResult

Tweaked class RuleCheck so that it will stream messages to a
printstream as well as collect them in theits RuleCheckResult

Modified TestScript class -- the one used by bmcheck -- so that it
delegates to class RuleCheck

Replaced RuleCheck.isOk() with RuleCheck.hasErrors() to be in keeping
with hasWarnings() and modified TestRuleCheck and RuleCheckMojo to use
this method to test the result.

Made TestRuleCheck print warnings as well as errors so we see at build
time which rules in the test rule set have warnigns against them.