Extra test cases to validate behaviour of fix for BYTEMAN-237
The test cases pin down the way RETURN and THROW work in AT EXCEPTION
EXIT rules, specifically how this manifests when mro ethan one AT
EXCEPTION EXIT rule is employed in the same method.

With no THROW or RETURN in the first rule an uncaught exception
passing through the target method is processed by each AT EXCEPTION
EXIT rule in turn (as exemplified by the call to method
testMethodMultiRule3).

When the first rule uses THROW the second rule is still triggered but
it is passed the exception THROWN by the first rule which need not be
the original exception (as exemplified by the call to method
testMethodMultiRule). That's contrary to existing documented behaviour
where a THROW is supposed to short-circuit further method and rule
execution with an exceptional return from the target method. This
behaviour needs special-casing in the user guide.

When the first rule uses RETURN (as exemplified by the call to method
testMethodMultiRule2) this bypasses the 2nd rule i.e. it
short-circuits further rule processing as per the existing documented
behaviour.

These behaviours are acceptable so long as they are properly
documented.