interpreter must rebox when initializer requires numeric conversion -- part of fix for BYTEMAN-316
when a numeric result is used to initialize a numeric BIND var
the initializaton may require a numeric type conversion. the
interpreter works only in the boxed realm so in such cases it
needs to to do a corresponding conversion from one boxed type
to another. this case was observed during fixing of BYTEMAN-316
and the fix here forms part of the fix for that issue.