Updating byteman sample ThreadHistoryMonitorHelper to store all events
Updating and refactoring ThreadHistoryMonitorHelper to get stored all
events which happens during program execution.

Till now Map was used with key based on thread name. Thread name can be easily
the same for several threads. History was quite sparse.

By my understanding thread event history should be a log that stores all
events and it's up to developer to query the results.
That's the reson for this change.