work round half-arsed 'fix' to jar plugin
the plugin now refuses to run extra jar goals without a classifier
because it has decided that will overwrite the default jar product
it does this even when configuration of a final name (not to mention
output dir) means that there is actually no clash between the default
goal target and the extra goal targets. sheer, incompetent, half-arsed
stupidity has won again. bah humbug and don't even mention Tiny Tim.

so as a workaround for this nonsense we have to generate the test
module jars using a "lib" classifier and reference them from the
module.xml using the suffix "-lib".