Fix BMUnit resource-based script lookup
The original lookup logic involved function `normalize` which is intended for file-system path manipulation. On Windows this means that all path components are separated by a backslash; however such paths are not usable in `Classloader.getResource()`. The logic was fixed to use `/` for resource path construction.

In addition, it appears that methods `getLoadDirectory` and `getResourceLoadDirectory` in `org.jboss.byteman.contrib.bmunit.BMUnitConfigState` did incorrectly determine whether to do a chained lookup, contradicting associated comments and possibly causing NullPointerException. The respective conditions were fixed.